serverless_app_l3/requirements.txt
aws-cdk-lib==2.160.0
constructs>=10.0.0,<11.0.0
aws_solutions_constructs.aws_lambda_dynamodb>=2.71.0